user/requirements.txt
-i https://pypi.org/simple
aiofiles==24.1.0; python_version >= '3.8'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.18; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
alembic==1.15.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
async-property==0.2.2
asyncpg==0.30.0; python_full_version >= '3.8.0'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
deprecation==2.1.0
dnspython==2.7.0; python_version >= '3.9'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
fastapi==0.115.12; python_version >= '3.8'
frozenlist==1.6.0; python_version >= '3.9'
greenlet==3.2.1; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
joblib==1.4.2; python_version >= '3.8'
jwcrypto==1.5.6; python_version >= '3.8'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
motor==3.7.0; python_version >= '3.9'
multidict==6.4.3; python_version >= '3.9'
nltk==3.9.1; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
propcache==0.3.1; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.4.8
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.4; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.9.1; python_version >= '3.9'
pymongo==4.12.1; python_version >= '3.9'
pytest==8.3.5; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
python-jose==3.4.0
python-keycloak==5.5.0; python_version >= '3.9' and python_version < '4.0'
python-multipart==0.0.20; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
ruff==0.11.8; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.40; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspection==0.4.0; python_version >= '3.9'
urllib3==2.4.0; python_version >= '3.9'
uvicorn==0.34.2; python_version >= '3.9'
yarl==1.20.0; python_version >= '3.9'
